+++ CHANGELOG.md
@@ -18,6 +18,8 @@ a release.
 ---
 
 ## [Unreleased]
+#### Added
+- Sluggable: Add support for DateTimeImmutable fields
 
 #### Added
 - Tree: [NestedSet] `childrenQueryBuilder()` to allow specifying sort order separately for each field